requirements.txt
imageio==2.9.0
mmcv==1.7.0
numpy>=1.20
opencv-python~=4.7.0.72
Pillow~=9.4.0
PyYAML~=6.0
scikit-learn~=1.2.2
scipy~=1.10.1
setuptools<81
tensorboardX~=2.6
torch
torchvision
tqdm~=4.51.0
